refactor: improve PoolType argument handling
